requirements.txt
blessed==1.20.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
idna==3.4
inquirer==3.1.3
python-dotenv==1.0.0
python-editor==1.0.4
readchar==4.0.5
requests==2.31.0
six==1.16.0
termcolor==2.3.0
typer==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
yaspin==3.0.0
